% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgh_helpers.R
\name{as.mgh}
\alias{as.mgh}
\title{Create an object structured like an MGH object}
\usage{
as.mgh(
  x,
  version = 1L,
  ndim1 = as.integer(length(x)),
  ndim2 = 1L,
  ndim3 = 1L,
  nframes = 1L,
  type = 3L,
  dof = 0L
)
}
\arguments{
\item{x}{The vertex-wise values}

\item{version}{Version (default = 1)}

\item{ndim1}{Width / 1st dimension (default = length(x))}

\item{ndim2}{Height / 2nd dimension (default = 1)}

\item{ndim3}{Depth / 3rd dimension (default = 1)}

\item{nframes}{Number of scalar components (default = 1)}

\item{type}{Data type, can be UCHAR (0), SHORT (4), INT (1) or FLOAT (3) (default = 3)}

\item{dof}{Degrees of freedom (default = 0)}
}
\value{
A mgh object with data and various header elements
}
\description{
Create an object structured like an MGH object
}
\author{
Sander Lamballais, 2018.
}
