% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vw_pool.R
\name{barnard.rubin}
\alias{barnard.rubin}
\title{Pooled degrees of freedom calculation}
\usage{
barnard.rubin(lambda, m, dfcom = Inf)
}
\arguments{
\item{lambda}{: Proportion of total variance due to missingness}

\item{m}{: Number of imputed datasets}

\item{dfcom}{: Residual degrees of freedom, estimated using \code{stats::df.residual()}}
}
\value{
Value for the residual degrees of freedom
}
\description{
Edited from \code{mice::barnard.rubin}. This function computes the residual
degrees of freedom for hypothesis testing, as proposed by Barnard & Rubin (1999).
It is used by the \code{\link{vw_pool}} function.
}
\references{
Barnard, J. and Rubin, D.B. (1999). Small sample degrees of
freedom with multiple imputation. \emph{Biometrika}, 86, 948-955.
}
