% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_helpers.R
\name{build_output_bks}
\alias{build_output_bks}
\title{Generate a set of backing file (bk) names}
\usage{
build_output_bks(result_path, res_bk_names, verbose = TRUE)
}
\description{
Generate a set of backing file (bk) names
}
\keyword{internal}
