% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_helpers.R
\name{build_output_fbm}
\alias{build_output_fbm}
\title{Generate a set of backing file (bk) names}
\usage{
build_output_fbm(
  result_path,
  type,
  nrow,
  ncol,
  precision = "float",
  verbose = FALSE
)
}
\description{
Generate a set of backing file (bk) names
}
\keyword{internal}
