% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_supersubject.R
\name{build_supersubject}
\alias{build_supersubject}
\title{Build "supersubject" by stacking all vertex data in one large file-backed
matrix with dimensions n_subjects x n_vertices.}
\usage{
build_supersubject(
  subj_dir,
  folder_ids,
  outp_dir,
  measure,
  hemi,
  fwhmc = "fwhm10",
  target = "fsaverage",
  backing = file.path(outp_dir, paste(hemi, measure, target, "supersubject.bk", sep =
    ".")),
  error_cutoff = 20,
  mask = TRUE,
  save_rds = FALSE,
  cluster,
  verbose = TRUE
)
}
\arguments{
\item{subj_dir}{: path to the FreeSurfer data, this expects a verywise structure.}

\item{folder_ids}{: the vector of observations to include. This holds the relative
path (from `subj_dir`) to the FreeSurfer data folder (e.g. "site1/sub-1_ses-01").}

\item{outp_dir}{: output path, where logs, backing files and the matrix itself
(if `save_rds == TRUE`) will be stored.}

\item{measure}{: vertex-wise measure, used to identify files.}

\item{hemi}{: hemisphere, used to identify files.}

\item{fwhmc}{: (default = "fwhm10") full-width half maximum value, used to identify files.}

\item{target}{: (default = "fsaverage"), used to identify files.}

\item{backing}{: (default = `outp_dir`) location to save the matrix \code{backingfile}.}

\item{error_cutoff}{: (default = 20) how many missing directories or brain surface files
for the function to stop with an error. If < `error_cutoff` directories/files are not
found a warning is thrown and missing files are registered in the `issues.log` file.}

\item{mask}{: (default = TRUE) only keep cortical vertices, according to FreeSurfer
cortical map.}

\item{save_rds}{: (default = FALSE) save the supersubject file metadata for re-use
in other sessions.}

\item{cluster}{: the parallel cluster}

\item{verbose}{: (default = TRUE)}
}
\value{
A Filebacked Big Matrix with vertex data for all subjects (dimensions:
n_subjects x n_vertices)
}
\description{
Build "supersubject" by stacking all vertex data in one large file-backed
matrix with dimensions n_subjects x n_vertices.
}
\author{
Serena Defina, 2024.
}
