% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenotype.R
\name{check_pheno_obj}
\alias{check_pheno_obj}
\title{Check that object exists in the global environment}
\usage{
check_pheno_obj(obj_name)
}
\arguments{
\item{obj_name}{A character string specifying the name of the object}
}
\value{
The object from the global environment (if it exists).
}
\description{
Check that object exists in the global environment
}
\details{
This function checks if the specified object exists in the global environment.
It returns the object if found, otherwise raises an error with a message.
}
