% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fed_helpers.R
\name{chunk_Ymats}
\alias{chunk_Ymats}
\title{Compute the sum and sum of squares of Y}
\usage{
chunk_Ymats(X, ss_chunk)
}
\arguments{
\item{X}{Numeric matrix. The design matrix of dimension \eqn{n \times p}.}

\item{ss_chunk}{Numeric matrix. A chunk of the ss matrix of dimension \eqn{n \times B},
where B is the block size (number of vertices in the chunk).}
}
\value{
A list containing:
\describe{
\item{XtY}{A numeric matrix of dimension \eqn{p \times B}, representing \eqn{X^T Y}.}
\item{psumsY}{A numeric matrix of dimension \eqn{2 \times B}. The first row contains
the column sums of Y (\eqn{1^T Y}), and the second row contains the column sums
of squared Y (\eqn{Y^T Y}).}
}
}
\description{
Calculates the cross-product of the design matrix X with the vertex chunk Y (XtY),
along with the sum (1'Y) and sum of squares (Y'Y) for each vertex in the chunk.
}
\author{
Serena Defina, 2026.
}
