% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_freesurfer.R
\name{compute_clusters}
\alias{compute_clusters}
\title{Compute significant clusters of vertices}
\usage{
compute_clusters(
  outp_dir,
  hemi,
  term_number,
  fwhm,
  FS_HOME,
  cwp_thr = 0.025,
  mcz_thr = 30,
  csd_sign = "abs",
  mask = NULL,
  verbose = FALSE
)
}
\arguments{
\item{outp_dir}{: output path, where results are stored.}

\item{hemi}{: hemisphere.}

\item{term_number}{: which (fixed) term are the clusters computed for.}

\item{fwhm}{: full-width half maximum estimate of data smoothness.}

\item{FS_HOME}{: FreeSurfer directory, i.e. \code{$FREESURFER_HOME}}

\item{cwp_thr}{: (default = 0.025, when both hemispheres are ran, else 0.05)
the cluster-wise p-value threshold on top of all corrections.}

\item{mcz_thr}{: (default = 0.001) numeric value for the Monte Carlo simulation threshold.
Any of the following are accepted (equivalent values separate by `/`):
 * 13 / 1.3 / 0.05,
 * 20 / 2.0 / 0.01,
. * 23 / 2.3 / 0.005,
 * 30 / 3.0 / 0.001, \* default
 * 33 / 3.3 / 0.0005,
 * 40 / 4.0 / 0.0001.}

\item{csd_sign}{: (default = "abs")}

\item{mask}{: apply a mask (default = cortical mask + excluding problematic vertices)}

\item{verbose}{: (default = FALSE) verbosity.}
}
\description{
This function is a wrapper for the FreeSurfer command \code{mri_surfcluster}.
It computes the clusters of significant vertices.
}
