% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R, R/vw_summarize.R
\name{.print_diff_cutoff}
\alias{.print_diff_cutoff}
\title{Print one cutoff line (or pair of lines) for a difference map}
\usage{
.print_diff_cutoff(all_diff, cutoff, label_a, label_b, n_finite)

.print_diff_cutoff(all_diff, cutoff, label_a, label_b, n_finite)
}
\description{
Print one cutoff line (or pair of lines) for a difference map

Print one cutoff line (or pair of lines) for a difference map
}
\keyword{internal}
