% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vw_summarize.R
\name{.print_diff_stats}
\alias{.print_diff_stats}
\title{Print difference map summary statistics}
\usage{
.print_diff_stats(all_diff, label_a, label_b, digits = 3, pad = 8)
}
\description{
Print difference map summary statistics
}
\keyword{internal}
