% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_freesurfer.R
\name{estimate_fwhm}
\alias{estimate_fwhm}
\title{Estimate full-width half maximum (FWHM)}
\usage{
estimate_fwhm(
  result_path,
  hemi,
  mask = NULL,
  fs_template = "fsaverage",
  verbose = FALSE
)
}
\arguments{
\item{result_path}{: path where model residuals are stored and where results
are generated: this includes two files:
"hemi.measure.fwhm.dat" and "hemi.measure.finamMask.mgh".}

\item{hemi}{: hemisphere.}

\item{mask}{: apply a mask (default = cortical mask + excluding problematic vertices)}

\item{fs_template}{: (default = "fsaverage") data template.}

\item{verbose}{: (default = FALSE) verbosity.}
}
\value{
An integer FWHM value
}
\description{
This function is a wrapper for the FreeSurfer command \code{mris_fwhm}. It
estimates the full-width half maximum (FWHM) of the brain data based on the
residuals of your model.
}
