% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm2mgh.R
\name{fbm2mgh}
\alias{fbm2mgh}
\title{Save file backed matrix (FBM) to MGH file}
\usage{
fbm2mgh(fbm, fname, filter = NULL)
}
\arguments{
\item{fbm}{a file backed matrix}

\item{fname}{file name to be used to save out the data}

\item{filter}{a vector of indices of rows to include of fbm}
}
\description{
Save file backed matrix (FBM) to MGH file
}
\author{
Sander Lamballais, 2018.
Modified version of save_mgh.m (by Heath Pardoe, 09/12/2013)
}
