% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fbm_col_has_0}
\alias{fbm_col_has_0}
\title{Check if all row elements are 0 in FBM}
\usage{
fbm_col_has_0(
  X,
  n_cores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL,
  verbose = TRUE
)
}
\arguments{
\item{X}{: the file-backed matrix (FBM) object}

\item{n_cores}{: (default = 1) number of cores for parellalization}

\item{row.ind}{: indicator for rows}

\item{col.ind}{: indicator for columns}

\item{row.mask}{: (default = NULL) specify a subset of rows}

\item{col.mask}{: (default = NULL) specify a subset of columns}

\item{verbose}{: (default = TRUE)}
}
\value{
A (large) logical vector for where rows are all 0.
}
\description{
Check if all row elements in a FBM are 0. This is used to clean up the
super-subject matrix in \code{\link{mask_cortex}}.
}
\author{
Serena Defina, 2024.
}
