% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_helpers.R
\name{fbm_col_has_na}
\alias{fbm_col_has_na}
\title{Check if any elements are NA in FBM}
\usage{
fbm_col_has_na(
  X,
  n_cores = 1,
  row.ind = bigstatsr::rows_along(X),
  col.ind = bigstatsr::cols_along(X),
  row.mask = NULL,
  col.mask = NULL,
  verbose = FALSE
)
}
\arguments{
\item{X}{: the file-backed matrix (FBM) object}

\item{n_cores}{: (default = 1) number of cores for parellalization}

\item{row.ind}{: indicator for rows}

\item{col.ind}{: indicator for columns}

\item{row.mask}{: (default = NULL) specify a subset of rows}

\item{col.mask}{: (default = NULL) specify a subset of columns}

\item{verbose}{: (default = TRUE)}
}
\value{
A (large) logical vector for where a column contained at least 2 observed values.
}
\description{
Check if any row elements in a FBM are NA. This is used to clean up the
effect sizes vectors before running a meta-analysis.
}
\author{
Serena Defina, 2026.
}
