% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp2list.R
\name{imp2list}
\alias{imp2list}
\title{Convert imputation object to a list of dataframes}
\usage{
imp2list(obj)
}
\arguments{
\item{obj}{The imputation object that contains the imputed datasets}
}
\value{
A list of imputed datasets
}
\description{
Converts known imputation objects (e.g., \code{amelia}, \code{aregImpute},
\code{mi}, \code{mids}, \code{missForest}) to a \code{list} of dataframes.
}
\details{
This function attempts to unify imputation object formats in supplying
datasets. This is done by extracting the imputed datasets from the imputation
object and assembling them into a \code{list}. If an unknown object is
supplied, the function will throw an error with the request to supply a list
of data frames instead.

This is a generic function: methods can be defined for it directly,
see \code{methods(imp2list)} for a list of all available methods.
}
\author{
Sander Lamballais, 2018.
}
