% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list.dirs.till}
\alias{list.dirs.till}
\title{List sub-directories till depth \code{n}}
\usage{
list.dirs.till(path, n)
}
\arguments{
\item{path}{: the path where to begin search}

\item{n}{: the depth of the recursive process (i.e., how many levels of sub-directories to list)}
}
\value{
A list of directory/sub-directory names.
}
\description{
Small recursive utility to list directories given complex folder structure.
This is used in \code{\link{build_supersubject}} for listing all sub-directories
until the correct level.
}
\author{
Serena Defina, 2024.
}
