% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgh_helpers.R
\name{load.mgh}
\alias{load.mgh}
\title{Load an MGH file into memory}
\usage{
load.mgh(file_path, as_vector = TRUE)
}
\arguments{
\item{file_path}{The (full) path to the mgh file.}

\item{as_vector}{Logical. Return the entire object or just the vertex values
as a vector.}
}
\value{
A mgh object with data and various header elements
}
\description{
Load an MGH file into memory
}
\author{
Heath Perdoe, 2013.
}
