% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenotype.R
\name{load_pheno_file}
\alias{load_pheno_file}
\title{Load "phenotype" file into R based on its extension}
\usage{
load_pheno_file(pheno_str, verbose = TRUE)
}
\arguments{
\item{pheno_str}{A character string specifying the file path.}

\item{verbose}{(default = TRUE).}
}
\value{
A data frame containing the loaded data
}
\description{
Takes a path to a file and reads it into memory.
}
\details{
The function checks if the file exists, identifies the file extension, and
loads the data accordingly. Supported file types: .csv, .rds, .sav, and .txt.
}
