% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vw_postprocess.R
\name{move_result_files}
\alias{move_result_files}
\title{Move key result files from one directory to another (for sharing and visualization)}
\usage{
move_result_files(from_dir, to_dir)
}
\arguments{
\item{from_dir}{Character string indicating the path to the "source" results
directory (i.e. \code{outp_dir} in the analysis call)}

\item{to_dir}{Character string indicating the path to the directory where
matching files will be copied.
Note: any required sub-directories will be created automatically.}
}
\value{
Invisibly returns a logical vector indicating whether each file was
successfully copied.
Side effects: \code{.mgh} files are copied to \code{to_dir}.
}
\description{
The function searches through a \code{verywise} results directory for all files
needed for visualization: i.e. clusters, coefficient maps, and "stack names".
It then copies them to a destination directory while preserving any
sub-directory structure in the source results folder.
}
\details{
Files are matched using these regular expression patterns:
\itemize{
\item \code{"[a-z]{1}h.[a-z]+.stack[0-9]{1,2}.cache.th30.abs.sig.ocn.mgh"}:
cluster files
\item \code{"[a-z]{1}h.[a-z]+.stack[0-9]{1,2}.coef.mgh"}:
coefficient (beta) maps
\item \code{"stack_names.txt"}:
stack names
}
}
\examples{
\dontrun{
# Move matching files from "results" to "plots"
move_result_files("path/to/results", "path/to/plots")
}

}
