% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vw_message.R
\name{prettify_message}
\alias{prettify_message}
\title{Print pretty messages to console}
\usage{
prettify_message(content, fill = "=", tot_nchar = 80, center = TRUE)
}
\arguments{
\item{content}{: The message to print}

\item{fill}{: (default "=")}

\item{tot_nchar}{: (default: 60)}

\item{center}{: (default: TRUE) center the content.}
}
\value{
A prettyfyed console message.
}
\description{
Adds fillers to messages
}
\author{
Serena Defina, 2025.
}
