% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_lmm2.R
\name{run_vw_lmm2}
\alias{run_vw_lmm2}
\title{Run vertex-wise linear mixed model using \code{\link[lme4:lmer]{lme4::lmer()}} (without refitting)}
\usage{
run_vw_lmm2(
  formula,
  pheno,
  subj_dir,
  outp_dir = NULL,
  hemi = c("lh", "rh"),
  fs_template = "fsaverage",
  apply_cortical_mask = TRUE,
  folder_id = "folder_id",
  tolerate_surf_not_found = 20,
  weights = NULL,
  lmm_control = lme4::lmerControl(calc.derivs = FALSE),
  REML = TRUE,
  seed = 3108,
  n_cores = 1,
  chunk_size = 1000,
  mtc = c("mcz", "fdr"),
  FS_HOME = Sys.getenv("FREESURFER_HOME"),
  fwhm = 10,
  mcz_thr = 30,
  cwp_thr = 0.025,
  save_optional_cluster_info = FALSE,
  save_ss = FALSE,
  save_residuals = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{formula}{A model formula object. This should specify a linear mixed
model \code{lme4} syntax. The outcome variable should be one of the
supported brain surface metrics (see Details). Example:
\code{vw_thickness ~ age * sex + site + (1|participant_id)}.}

\item{pheno}{Either a \code{data.frame}/\code{tibble} containing the
"phenotype" data (i.e., already loaded in the global environment), or a
string specifying the file path to phenotype data. Supported file formats:
.rds, .csv, .txt, .sav (SPSS).
The data should be in \strong{long} format and it should contain all the
variables specified in the left-hand side of the \code{formula} (i.e., after the \code{~})
plus the \code{folder_id} column.}

\item{subj_dir}{Character string specifying the path to FreeSurfer data
directory. Must follow the verywise directory structure (see package
vignette for details).}

\item{outp_dir}{Character string specifying the output directory for results.
If \code{NULL} (default), creates a "verywise_results" sub-directory in the
current working directory (not recommended).}

\item{hemi}{Character string specifying which hemisphere to analyze.
Options: \code{"lh"} (left hemisphere: default), \code{"rh"} (right hemisphere).}

\item{fs_template}{Character string specifying the FreeSurfer template for
vertex registration. Options:
\itemize{
\item \code{"fsaverage"} (default) = 163842 vertices (highest resolution),
\item \code{"fsaverage6"} = 40962 vertices,
\item \code{"fsaverage5"} = 10242 vertices,
\item \code{"fsaverage4"} = 2562 vertices,
\item \code{"fsaverage3"} = 642 vertices
}
Note that lower resolutions should be only used to downsample the brain
map, for faster model tuning. The final analyses should also run using
\code{fs_template = "fsaverage"} to avoid (small) imprecisions in vertex
registration and smoothing.}

\item{apply_cortical_mask}{Logical indicating whether to exclude non-cortical
vertices from analysis. Default: \code{TRUE} (recommended).}

\item{folder_id}{Character string specifying the column name in \code{pheno}
that contains subject directory names of the input neuroimaging data
(e.g. "sub-001_ses-baseline" or "site1/sub-010_ses-F1"). These are expected
to be nested inside \code{subj_dir}.
Default: \code{"folder_id"}.}

\item{tolerate_surf_not_found}{Integer indicating how many brain surface
files listed in \code{folder_id} can be missing from \code{subj_dir}. If
the number of missing or corrupted files is
\code{ > tolerate_surf_not_found } execution will stop.
Default: \code{20L}.}

\item{weights}{Optional string or numeric vector of weights for the linear mixed model.
You can use this argument to specify inverse-probability weights. If this
is a string, the function look for a column with that name in the phenotype
data. Note that these are not normalized or standardized in any way.
Default: \code{NULL} (no weights).}

\item{lmm_control}{Optional list (of correct class, resulting from
\code{lmerControl()} containing control parameters to be passed to
\code{lme4::lmer()} (e.g. optimizer choice, convergence criteria,
see the \code{?lmerControl} documentation for details.
Default: (uses default settings).}

\item{REML}{Logical specifying whether to optimize the REML criterion (as opposed to
the log-likelihood). Default: TRUE. Use \code{REML = FALSE} if you intend to do model
comparison (using AIC output).}

\item{seed}{Integer specifying the random seed for reproducibility
Default: 3108.}

\item{n_cores}{Integer specifying the number of CPU cores for parallel
processing.
Default: 1.}

\item{chunk_size}{Integer specifying the number of vertices processed per
chunk in parallel operations. Larger values use more memory but may be
faster.
Default: 1000.}

\item{mtc}{Character string: multiple testing correction strategy.
Options: \code{"mcz"} (FreeSurfer MonteCarlo-based cluster correction: default), \code{"fdr"} (False Discovery Rate).}

\item{FS_HOME}{Character string specifying the FreeSurfer home directory.
Defaults to \code{FREESURFER_HOME} environment variable.}

\item{fwhm}{Numeric value specifying the full-width half-maximum for
smoothing kernel. Default: 10.}

\item{mcz_thr}{Numeric value for the Monte Carlo simulation threshold.
Any of the following are accepted (equivalent values are separated by \code{/}):
\itemize{
\item 13 / 1.3 / 0.05,
\item 20 / 2.0 / 0.01,
\item 23 / 2.3 / 0.005,
\item 30 / 3.0 / 0.001, \* default
\item 33 / 3.3 / 0.0005,
\item 40 / 4.0 / 0.0001.
}}

\item{cwp_thr}{Numeric value for cluster-wise p-value threshold (on top of all
corrections). Set this should be set to \code{0.025} when both hemispheres are analyzed,
and \code{0.05} for single hemisphere analyses.
Default: \code{0.025}.}

\item{save_optional_cluster_info}{Logical indicating whether to save additional
output form \code{mri_surfcluster} call. See \code{\link[=compute_clusters]{compute_clusters()}}
for details. Default: \code{FALSE}.}

\item{save_ss}{Logical indicating whether to save the super-subject matrix
("ss") as an .rds file that can be then re-used in future analyses. This can
also be a character string specifying the directory where ss should be saved.
When \code{TRUE}, the ss matrix will be saved in \verb{<outp_dir>/ss} by
default. Default: \code{FALSE}.}

\item{save_residuals}{Logical indicating whether to save the residuals.mgh
file. Default: \code{FALSE}.}

\item{verbose}{Logical indicating whether to display progress messages.
Default: \code{TRUE}.}
}
\description{
This is an alternative to the main function for conducting vertex-wise
linear mixed model analyses on brain surface metrics.
This is almost identical to \code{\link[=run_vw_lmm]{run_vw_lmm()}} but it estimates a linear
mixed model from scratch at each vertex (instead of refitting from a template)
This is done using the \code{\link[=single_lmm]{single_lmm()}} function.

Use this pipeline when you have reason to believe that the random effect
structure is very different across vertices. In all test I have run so far,
\code{run_vw_lmm()} and \code{run_vw_lmm2()} give identical fixed term results.
}
\details{
See \code{\link[=run_vw_lmm]{run_vw_lmm()}} for details.
}
\author{
Serena Defina, 2026.
}
