% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgh_helpers.R
\name{save.mgh}
\alias{save.mgh}
\title{Save an MGH file from memory}
\usage{
save.mgh(vol, file_name)
}
\arguments{
\item{vol}{The MGH object (as from load.mgh)}

\item{file_name}{File name to be used to save out the data}
}
\description{
R translation of save_mgh.m (by Heath Pardoe, 09/12/2013)
}
\author{
Sander Lamballais, 2018.
}
