% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataset.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Simulate a dataset including phenotype and FreeSurfer data}
\usage{
simulate_dataset(
  path,
  data_structure = list(cohort1 = list(sessions = c("01", "02", "03"), n_subjects = 10),
    cohort2 = list(sessions = c("01", "02"), n_subjects = 20)),
  seed = 31081996,
  ...
)
}
\arguments{
\item{path}{: Where should the dataset be created.}

\item{data_structure}{: A nested list, with top level determining the cohort/
dataset/site. Each site is itself a list with two items: \code{"sessions"}: a
vector of session names/numbers; and \code{"n_subjects"}: aninteger indicating
the number of subjects.}

\item{seed}{: (default = 31081996) seed used for randomization.}

\item{...}{: Other arguments to be passed to \code{\link{simulate_freesurfer_data}}}
}
\description{
Simulating a dataset with multiple cohorts/sites and multiple
timepoints/sessions.
}
