% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_helpers.R
\name{unpack_formula}
\alias{unpack_formula}
\title{Unpack R formula}
\usage{
unpack_formula(formula, df, return_X = FALSE)
}
\arguments{
\item{formula}{A formula object, for example specifying a LME model.}

\item{df}{A data.frame, for example the first element in the list
output of \code{\link{imp2list}}.}

\item{return_X}{Logical. Whether to return the whole design matrix or
just the (fixed) term names.}
}
\value{
A design matrix or a character vector of (fixed) term names.
}
\description{
Get the terms for the (fixed) effect for the given formula.
Wors for \code{lme4} as well as regular \code{lm} style formulas.
}
\author{
Serena Defina, 2026.
}
