% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fed_helpers.R
\name{vw_chol_solve}
\alias{vw_chol_solve}
\title{Solve a linear system given a Cholesky factor}
\usage{
vw_chol_solve(R, B)
}
\arguments{
\item{R}{Upper-triangular matrix. Cholesky factor of \eqn{A} as returned by
\code{\link{vw_chol}} or \code{\link[base]{chol}}.}

\item{B}{Numeric matrix (or vector). Right-hand side; must have the same
number of rows as \code{R}.}
}
\value{
Numeric matrix \eqn{X = A^{-1} B} of the same dimensions as \code{B}.
}
\description{
Solves \eqn{A X = B} for \eqn{X} given the upper-triangular Cholesky factor
\eqn{R} of \eqn{A} (i.e. \eqn{R^\top R = A}), using forward and backward
substitution. This is numerically preferable to forming \eqn{A^{-1}} explicitly.
}
\keyword{internal}
