% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vw_message.R
\name{vw_init_message}
\alias{vw_init_message}
\title{Print pipeline title with verywise version noted}
\usage{
vw_init_message(model_type, verbose = TRUE)
}
\description{
Print pipeline title with verywise version noted
}
\keyword{internal}
