% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vw_message.R
\name{vw_pretty_message}
\alias{vw_pretty_message}
\title{Print decorated message to console if verbose = TRUE}
\usage{
vw_pretty_message(msg, ..., verbose = TRUE)
}
\arguments{
\item{msg}{Content to print}

\item{...}{Other arguments to pass to prettify_message}

\item{verbose}{(default = TRUE)}
}
\description{
Print decorated message to console if verbose = TRUE
}
