% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_supersubject.R
\name{build_supersubject}
\alias{build_supersubject}
\title{Build "supersubject" by stacking all vertex data in one large file-backed
matrix with dimensions n_subjects x n_vertices.}
\usage{
build_supersubject(
  subj_dir,
  folder_id,
  files_list = list.dirs.till(subj_dir, n = 2),
  measure = "thickness",
  hemi = "lh",
  fwhmc = "fwhm10",
  target = "fsaverage",
  backing = file.path(subj_dir, paste0(hemi, ".", measure, ".supersubject.bk")),
  mask = TRUE,
  save_rds = FALSE,
  cluster,
  verbose = TRUE
)
}
\arguments{
\item{subj_dir}{: path to the FreeSurfer data, this expects a verywise structure.}

\item{folder_id}{: column in phenotype file that holds the IDs of the observations
to include. This also expects a verywise format (e.g. "site1/sub-1_ses-01")}

\item{files_list}{: (default = \code{\link{list.dirs.till}(subj_dir, n = 2)})
allows to specify manually which folders or files to include. You would typically
not use this parameter, but rather use \code{folder_id} to specify which observations
to include.}

\item{measure}{: (default = "thickness"), vertex-wise measure, used to identify files.}

\item{hemi}{: (default = "lh") hemisphere, used to identify files.}

\item{fwhmc}{: (default = "fwhm10") full-width half maximum value, used to identify files.}

\item{target}{: (default = "fsaverage"), used to identify files.}

\item{backing}{: (default = \code{subj_dir}) location to save the matrix \code{backingfile}.}

\item{mask}{: (dafault = TRUE) only keep cortical vertices, according to FreeSurfer
cortical map.}

\item{save_rds}{: (default = FALSE) save the supersubject file metadata for re-use
in other sessions.}

\item{cluster}{: the parallel cluster}

\item{verbose}{: (default = TRUE)}
}
\value{
A Filebacked Big Matrix with vertex data for all subjects (dimensions:
n_subjects x n_vertices)
}
\description{
Build "supersubject" by stacking all vertex data in one large file-backed
matrix with dimensions n_subjects x n_vertices.
}
\author{
Serena Defina, 2024.
}
