% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_supersubject.R
\name{build_supersubject}
\alias{build_supersubject}
\title{Build "supersubject" by stacking all vertex data in one matrix}
\usage{
build_supersubject(
  subj_dir,
  folder_id,
  files_list = list.dirs.till(subj_dir, n = 2),
  measure = "thickness",
  hemi = "lh",
  fwhmc = "fwhm10",
  target = "fsaverage",
  backing = file.path(subj_dir, paste0(hemi, ".", measure, "supersubject")),
  n_cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{subj_dir}{: path to the FreeSurfer data, expect verywise structure}

\item{folder_id}{: column in phenotype file that holds ids to include in
verywise folder format (e.g. "site1/sub-1_ses-01")}

\item{files_list}{: (default = \code{\link{list.dirs.till}(subj_dir, n = 2)})
but allows to specify manually.}

\item{measure}{: (default = "thickness"), measure, used to identify files.}

\item{hemi}{: (default = "lh") hemisphere, used to identify files.}

\item{fwhmc}{: (default = "fwhm10") full-width half maximum value, used to identify files.}

\item{target}{: (default = "fsaverage"), used to identify files.}

\item{backing}{: (default = \code{subj_dir}) location to save the matrix \code{backingfile}.}

\item{n_cores}{: (default = 1) number of cores for parallel processing.}

\item{verbose}{: (default = TRUE)}
}
\value{
A Filebacked Big Matrix with vertex data for all subjects (dimensions:
n_vertex * n_subject)
}
\description{
Build "supersubject" by stacking all vertex data in one matrix
}
\author{
Serena Defina, 2024.
}
