% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vw_postprocess.R
\name{convert_to_mgh}
\alias{convert_to_mgh}
\title{Convert statistical result FBMs to FreeSurfer \code{.mgh} format}
\usage{
convert_to_mgh(
  vw_results,
  result_path,
  fixed_terms = NULL,
  random_terms = NULL,
  stat_names = c("coef", "se", "p", "-log10p", "resid"),
  save_resid = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{vw_results}{A named list of \link[bigstatsr:FBM-class]{bigstatsr::FBM} objects containing the
statistical results.}

\item{result_path}{Character string indicating the base output path where the
\code{.mgh} files will be written.}

\item{fixed_terms}{Vector of fixed term names to be included in output filenames
(as "stack1", "stack2"...).}

\item{random_terms}{Vector of random term names used to label ICC statistics.}

\item{stat_names}{Character vector of statistic names to process.
Default: \code{c("coef", "se","p", "-log10p","resid")}.
The special names: \code{"-log10p"} and \code{"fdr"} trigger the on-the-fly p-value
transformations.}

\item{save_resid}{Logical. Keep the backing around to then save residuals as fbm.
Default: \code{FALSE}.}

\item{verbose}{Logical. Default:\code{TRUE}}
}
\value{
Invisibly returns \code{NULL}. Side effects: \code{.mgh} files are written to disk.
}
\description{
This function takes a list of \link[bigstatsr:FBM-class]{bigstatsr::FBM} objects storing statistical
results and writes them to FreeSurfer-compatible \code{.mgh} files.
It supports coefficient, standard error, p-value, and residual maps, with
optional on-the-fly \eqn{-\log_{10}} transformation of p-values.
}
\details{
For residuals, all rows are written into a single \code{.mgh} file. For other
stats, data is written to one file per row (i.e. term).
Note, by default, the p vector is cast down to float (single precision, 32‑bit),
meaning about 7 significant decimal digits are stored (accurately). This is done
to cut disk space needed for the results in half.

When computing \eqn{-\log_{10}(p)} values, the transformation is applied
\strong{in chunks of columns} to avoid loading the full FBM into memory.
}
