% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgh_helpers.R
\name{fbm2mgh}
\alias{fbm2mgh}
\title{Save file backed matrix (FBM) to MGH file}
\usage{
fbm2mgh(fbm, fnames, mode = "1row.1file")
}
\arguments{
\item{fbm}{a file backed matrix}

\item{fnames}{file name to be used to save out the data}

\item{mode}{"1row.1file" or "allrows.1file"}
}
\description{
Save file backed matrix (FBM) to MGH file
}
\author{
Serena Defina, 2025.
Modified version of save_mgh.m (by Heath Pardoe, 09/12/2013)
}
