% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_lmm.R
\name{hemi_vw_lmm}
\alias{hemi_vw_lmm}
\title{Run vertex-wise linear mixed model in one hemisphere using \code{lme4::lmer()}}
\usage{
hemi_vw_lmm(
  formula,
  subj_dir,
  data_list,
  hemi,
  measure = gsub("vw_", "", all.vars(formula)[1]),
  fwhm = 10,
  target = "fsaverage",
  seed = 3108,
  apply_cortical_mask = TRUE,
  model = "lme4::lmer"
)
}
\arguments{
\item{formula}{: model formula object (this should specify a LME model)}

\item{subj_dir}{: path to the FreeSurfer data, this expects a verywise structure.}

\item{data_list}{: the phenotype as a list of dataframes (generated by \code{\link{imp2list}})}

\item{hemi}{: (default = "both") hemispheres to run.}

\item{measure}{: (default = \code{gsub("vw_", "", all.vars(formula)[1])}) vertex-wise measure
this should be the same as specified in formula}

\item{fwhm}{: (default = 10) full-width half maximum value}

\item{target}{: (default = "fsaverage") template on which to register vertex-wise data.}

\item{seed}{: (default = 3108) random seed.}

\item{apply_cortical_mask}{: (default = TRUE) remove vertices that are not on the cortex.}

\item{model}{: (default = \code{"lme4::lmer"}) # "stats::lm"}
}
\value{
A list of file-backed matrices containing pooled coefficients, SEs,
t- and p- values and residuals.
}
\description{
It runs the \code{\link{single_lmm}} function across vertices in a single hemisphere.
}
\author{
Serena Defina, 2024.
}
