% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_lmm.R
\name{hemi_vw_lmm}
\alias{hemi_vw_lmm}
\title{Run vertex-wise linear mixed model in one hemisphere using \code{lme4::lmer()}}
\usage{
hemi_vw_lmm(
  formula,
  data_list,
  subj_dir,
  outp_dir = NULL,
  FS_HOME = "",
  folder_id = "folder_id",
  hemi,
  fwhm = 10,
  fs_template = "fsaverage",
  mcz_thr = 30,
  cwp_thr = 0.025,
  seed = 3108,
  apply_cortical_mask = TRUE,
  save_ss = TRUE,
  model = "lme4::lmer",
  use_model_template = TRUE,
  n_cores,
  chunk_size = 1000,
  verbose = TRUE
)
}
\arguments{
\item{formula}{: a string or model formula object specifying a LME model,
e.g. \code{vw_thickness ~ sex * age + site + (1|id)}.}

\item{data_list}{: a list of dataframes containing phenotype information
(generated by \code{\link{imp2list}})}

\item{subj_dir}{: a string containing the path to the neuroimaging (FreeSurfer
pre-processed) data, this expects a verywise structure.}

\item{outp_dir}{: output path, where do you want results to be stored. If none is
provided by the user, a "results" sub-directory will created inside \code{subj_dir}.}

\item{FS_HOME}{: FreeSurfer directory, i.e. \code{$FREESURFER_HOME}.}

\item{folder_id}{: (default = "folder_id") the name of the column in pheno that
contains the directory names of the input neuroimaging data (e.g. "sub-10_ses-T1").}

\item{hemi}{: (default = "both") hemispheres to run.}

\item{fwhm}{: (default = 10) full-width half maximum value}

\item{fs_template}{: (default = "fsaverage") template on which to register vertex-wise data.
The following values are accepted:
 * fsaverage (default) = 163842 vertices (highest resolution),
 * fsaverage6 = 40962 vertices,
 * fsaverage5 = 10242 vertices,
 * fsaverage4 = 2562 vertices,
 * fsaverage3 = 642 vertices
Note that, at the moment, these are only used to downsample the brain map, for faster
model tuning. `verywise` expects the input data to be always registered on the "fsaverage"
template and the final analyses should also be run using `fs_template = "fsaverage"`
to avoid (small) imprecisions in vertex registration and smoothing.}

\item{mcz_thr}{: (default = 0.001) numeric value for the Monte Carlo simulation threshold.
Any of the following are accepted (equivalent values separate by `/`):
 * 13 / 1.3 / 0.05,
 * 20 / 2.0 / 0.01,
 * 23 / 2.3 / 0.005,
 * 30 / 3.0 / 0.001, \* default
 * 33 / 3.3 / 0.0005,
 * 40 / 4.0 / 0.0001.}

\item{cwp_thr}{: (default = 0.025, when both hemispheres are ran, else 0.05)
the cluster-wise p-value threshold on top of all corrections.}

\item{seed}{: (default = 3108) random seed.}

\item{apply_cortical_mask}{: (default = TRUE) remove vertices that are not on the cortex.}

\item{save_ss}{: (default = TRUE) save the super-subject matrix as an rds file for
quicker processing in the future.}

\item{model}{: (default = \code{"lme4::lmer"}) # "stats::lm"}

\item{use_model_template}{: pre-compile the model?}

\item{n_cores}{: the parallel cluster}

\item{chunk_size}{: (default = 1000) size of data chunk for parallel processing}

\item{verbose}{: (default = TRUE)}
}
\value{
A list of file-backed matrices containing pooled coefficients, SEs,
t- and p- values and residuals.
}
\description{
It runs the \code{\link{single_lmm}} function across vertices in a single hemisphere.
}
\author{
Serena Defina, 2024.
}
