% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load.mgh}
\alias{load.mgh}
\title{Load an MGH file into memory}
\usage{
load.mgh(file_path)
}
\arguments{
\item{file_path}{: the (full) path to the mgh file.}
}
\value{
A mgh object with data and various header elements
}
\description{
Load an MGH file into memory
}
\author{
Heath Perdoe, 2013.
}
