% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_chunk_sequence}
\alias{make_chunk_sequence}
\title{Define chunks of vertices for analyses}
\usage{
make_chunk_sequence(iv, chunk_size = 1000)
}
\arguments{
\item{iv}{: independent variable, i.e. the vertex-wise brain data to use.}

\item{chunk_size}{: (default = 1000) how big are the chunks}
}
\value{
A list of chunk start and end positions.
}
\description{
This function takes a vector of values representing the dimension of the
vertex-wise data to use (i.e. the output of \code{\link{mask_cortex}}) and
splits it into chunks for speeding up parallel processing.
}
\author{
Serena Defina, 2024.
}
