% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_chunk_sequence}
\alias{make_chunk_sequence}
\title{Define chunks of vertices for analyses}
\usage{
make_chunk_sequence(iv, chunk_size = 1000)
}
\arguments{
\item{iv}{: indices of the vertex-wise brain data to use.}

\item{chunk_size}{: (default = 1000) how big are the chunks}
}
\value{
A list with n_chunks elements. Each element is a vector of vertex
  positions.
}
\description{
This function takes a vector of values representing the dimension of the
vertex-wise data to use and splits it into chunks for memory-efficient
parallel processing.
}
\author{
Serena Defina, 2024.
}
