% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_lmm.R
\name{run_vw_lmm}
\alias{run_vw_lmm}
\title{Run vertex-wise linear mixed model using \code{\link[lme4:lmer]{lme4::lmer()}}}
\usage{
run_vw_lmm(
  formula,
  pheno,
  subj_dir,
  outp_dir = NULL,
  hemi = c("lh", "rh"),
  fs_template = "fsaverage",
  apply_cortical_mask = TRUE,
  folder_id = "folder_id",
  tolerate_surf_not_found = 20,
  weights = NULL,
  lmm_control = lme4::lmerControl(calc.derivs = FALSE),
  REML = TRUE,
  seed = 3108,
  n_cores = 1,
  chunk_size = 1000,
  mtc = c("mcz", "fdr"),
  FS_HOME = Sys.getenv("FREESURFER_HOME"),
  fwhm = 10,
  mcz_thr = 30,
  cwp_thr = 0.025,
  save_optional_cluster_info = FALSE,
  save_ss = FALSE,
  save_residuals = FALSE,
  save_cov = NULL,
  verbose = TRUE
)
}
\arguments{
\item{formula}{A model formula object. This should specify a linear mixed
model \code{lme4} syntax. The outcome variable should be one of the
supported brain surface metrics (see Details). Example:
\code{vw_thickness ~ age * sex + site + (1|participant_id)}.}

\item{pheno}{Either a \code{data.frame}/\code{tibble} containing the
"phenotype" data (i.e., already loaded in the global environment), or a
string specifying the file path to phenotype data. Supported file formats:
.rds, .csv, .txt, .sav (SPSS).
The data should be in \strong{long} format and it should contain all the
variables specified in the left-hand side of the \code{formula} (i.e., after the \code{~})
plus the \code{folder_id} column.}

\item{subj_dir}{Character string specifying the path to FreeSurfer data
directory. Must follow the verywise directory structure (see package
vignette for details).}

\item{outp_dir}{Character string specifying the output directory for results.
If \code{NULL} (default), creates a "verywise_results" sub-directory in the
current working directory (not recommended).}

\item{hemi}{Character string specifying which hemisphere to analyze.
Options: \code{"lh"} (left hemisphere: default), \code{"rh"} (right hemisphere).}

\item{fs_template}{Character string specifying the FreeSurfer template for
vertex registration. Options:
\itemize{
\item \code{"fsaverage"} (default) = 163842 vertices (highest resolution),
\item \code{"fsaverage6"} = 40962 vertices,
\item \code{"fsaverage5"} = 10242 vertices,
\item \code{"fsaverage4"} = 2562 vertices,
\item \code{"fsaverage3"} = 642 vertices
}
Note that lower resolutions should be only used to downsample the brain
map, for faster model tuning. The final analyses should also run using
\code{fs_template = "fsaverage"} to avoid (small) imprecisions in vertex
registration and smoothing.}

\item{apply_cortical_mask}{Logical indicating whether to exclude non-cortical
vertices from analysis. Default: \code{TRUE} (recommended).}

\item{folder_id}{Character string specifying the column name in \code{pheno}
that contains subject directory names of the input neuroimaging data
(e.g. "sub-001_ses-baseline" or "site1/sub-010_ses-F1"). These are expected
to be nested inside \code{subj_dir}.
Default: \code{"folder_id"}.}

\item{tolerate_surf_not_found}{Integer indicating how many brain surface
files listed in \code{folder_id} can be missing from \code{subj_dir}. If
the number of missing or corrupted files is \verb{> tolerate_surf_not_found}
execution will stop.
Default: \code{20L}.}

\item{weights}{Optional string or numeric vector of weights for the linear mixed model.
You can use this argument to specify inverse-probability weights. If this
is a string, the function look for a column with that name in the phenotype
data. Note that these are not normalized or standardized in any way.
Default: \code{NULL} (no weights).}

\item{lmm_control}{Optional list (of correct class, resulting from
\code{\link[lme4:lmerControl]{lme4::lmerControl()}} containing control parameters to be passed to
\code{\link[lme4:lmer]{lme4::lmer()}} (e.g. optimizer choice, convergence criteria,
see the \code{?lmerControl} documentation for details.
Default: (uses default settings).}

\item{REML}{Logical specifying whether to optimize the REML criterion (as opposed to
the log-likelihood). Default: TRUE. Use \code{REML = FALSE} if you intend to do model
comparison (using AIC output).}

\item{seed}{Integer specifying the random seed for reproducibility
Default: 3108.}

\item{n_cores}{Integer specifying the number of CPU cores for parallel
processing.
Default: 1.}

\item{chunk_size}{Integer specifying the number of vertices processed per
chunk in parallel operations. Larger values use more memory but may be
faster.
Default: 1000.}

\item{mtc}{Character string: multiple testing correction strategy.
Options: \code{"mcz"} (FreeSurfer MonteCarlo-based cluster correction: default), \code{"fdr"} (False Discovery Rate).}

\item{FS_HOME}{Character string specifying the FreeSurfer home directory.
Defaults to \code{FREESURFER_HOME} environment variable.}

\item{fwhm}{Numeric value specifying the full-width half-maximum for
smoothing kernel. Default: 10.}

\item{mcz_thr}{Numeric value for the Monte Carlo simulation threshold.
Any of the following are accepted (equivalent values are separated by \code{/}):
\itemize{
\item 13 / 1.3 / 0.05,
\item 20 / 2.0 / 0.01,
\item 23 / 2.3 / 0.005,
\item 30 / 3.0 / 0.001, \* default
\item 33 / 3.3 / 0.0005,
\item 40 / 4.0 / 0.0001.
}}

\item{cwp_thr}{Numeric value for cluster-wise p-value threshold (on top of all
corrections). Set this should be set to \code{0.025} when both hemispheres are analyzed,
and \code{0.05} for single hemisphere analyses.
Default: \code{0.025}.}

\item{save_optional_cluster_info}{Logical indicating whether to save additional
output form \code{mri_surfcluster} call. See \code{\link[=compute_clusters]{compute_clusters()}}
for details. Default: \code{FALSE}.}

\item{save_ss}{Logical indicating whether to save the super-subject matrix
("ss") as an .rds file that can be then re-used in future analyses. This can
also be a character string specifying the directory where ss should be saved.
When \code{TRUE}, the ss matrix will be saved in \verb{<outp_dir>/ss} by
default. Default: \code{FALSE}.}

\item{save_residuals}{Logical indicating whether to save the residuals.mgh
file. Default: \code{FALSE}.}

\item{save_cov}{Character vector of two fixed effect term names of \code{NULL}.
When specified, the covariance between the two terms is extracted and saved
for later analysis (e.g. simple slopes).}

\item{verbose}{Logical indicating whether to display progress messages.
Default: \code{TRUE}.}
}
\value{
A list of file-backed matrices (\link[bigstatsr:FBM-class]{bigstatsr::FBM} objects)
containing pooled coefficients, SEs, t- and p- values and residuals.
Results are also automatically saved to disk in .mgh format.
}
\description{
This is the main function for conducting vertex-wise linear mixed model
analyses on brain surface metrics. It will first check use inputs, prepare
the phenotype data(list) and the brain (outcome) data, then it will fit a
linear mixed model at each vertex of the specified hemisphere using the
\code{\link[=refit_lmm]{refit_lmm()}} function, extract key statistics and perform multiple testing
correction.

The function supports analysis of both single and multiple imputed datasets.
It also automatically handles cortical masking, and provides two multiple
testing strategies: FDR or cluster-wise correction using FreeSurfer's Monte
Carlo simulation approach.
}
\details{
\strong{Supported Brain Surface Metrics:}
The outcome specified in \code{formula} should be a brain surface metric
among:
\itemize{
\item \code{vw_thickness} - Cortical thickness
\item \code{vw_area} - Cortical surface area (white  surface)
\item \code{vw_area.pial} - Cortical surface area (pial surface)
\item \code{vw_curv} - Mean curvature
\item \code{vw_jacobian_white} - Jacobian determinant (white surface)
\item \code{vw_pial} - Pial surface coordinates
\item \code{vw_pial_lgi} - Local gyrification index (pial surface)
\item \code{vw_sulc} - Sulcal depth
\item \code{vw_volume} - Gray matter volume
\item \code{vw_w_g.pct} - White/gray matter intensity ratio
\item \code{vw_white.H} - Mean curvature (white surface)
\item \code{vw_white.K} - Gaussian curvature (white surface)
}

\strong{Statistical Approach:}
The function uses \code{\link[lme4:lmer]{lme4::lmer()}} for mixed-effects modeling, enabling
analysis of longitudinal and hierarchical data. P-values are computed using
the t-as-z approximation, with cluster-wise correction applied using
FreeSurfer's Monte Carlo simulation approach.

\strong{Multiple Imputation:}
The function automatically detects and handles multiple imputed datasets
(created with \code{mice} or similar packages), pooling results according
to Rubin's rules.

\strong{Parallel processing:}
The \code{verywise} package employs a carefully designed parallelization
strategy to maximize computational efficiency while avoiding the
performance penalties associated with nested parallelization.
Left and right cortical hemispheres are processed sequentially by default.
Parallel processing of the two hemispheres (and/or different metrics, models)
should be handled by the user (e.g., using SLURM job arrays or similar,
see vignette on parallelization).
Within each hemisphere, vertices are divided into chunks of size
\code{chunk_size} and processed in parallel across \code{n_cores} workers
(when \code{n_cores > 1}). When multiple imputed datasets are present,
these are processed sequentially within each vertex.

Note that, on some systems, implicit parallelism in low-level matrix algebra
libraries (BLAS/LAPACK) can interfere with explicit parallelization. If you
feel like processing is taking too long, I recommend disabling these implicit
threading libraries before starting R.
For example:
\preformatted{
export OPENBLAS_NUM_THREADS=1
export OMP_NUM_THREADS=1
export MKL_NUM_THREADS=1
export VECLIB_MAXIMUM_THREADS=1
export NUMEXPR_NUM_THREADS=1
}

Also note that using a very large number of cores (e.g. >120) may sometimes
cause worker initialization or other issues (e.g. R parallel processes
limits)

\strong{Output Files:}
Results are saved in FreeSurfer-compatible .mgh format for visualization
using \code{verywise} plotting functions (\code{\link[=plot_vw_map]{plot_vw_map()}},\code{\link[=plot_vw_diff]{plot_vw_diff()}}...),
the \href{https://github.com/SereDef/verywise-wizard}{verywiseWIZard},
FreeView or other neuroimaging software.
}
\note{
\itemize{
\item Ensure FreeSurfer is properly installed and the
\code{FREESURFER_HOME} environment variable is set.
\item Large datasets may require substantial memory. Consider adjusting
\code{chunk_size} and \code{n_cores} based on your system specifications.
\item For reproducibility, always specify a \code{seed}.
}
}
\seealso{
\code{\link[=single_lmm]{single_lmm()}} for single-vertex modeling,
\code{vignette("03-run-vw-lmm", package = "verywise")} for detailed
usage examples.
}
\author{
Serena Defina, 2024.
}
