% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_lmm.R
\name{run_vw_lmm}
\alias{run_vw_lmm}
\title{Run vertex-wise linear mixed model using \code{lme4::lmer()}}
\usage{
run_vw_lmm(
  formula,
  pheno = NULL,
  subj_dir,
  hemi = "lh",
  measure = gsub("vw_", "", all.vars(formula)[1]),
  fwhm = 10,
  target = "fsaverage",
  n_cores = 1,
  model = "lme4::lmer"
)
}
\arguments{
\item{formula}{: model formula object (this should specify a LME model)}

\item{pheno}{: the phenotype data or a path to the data file.}

\item{subj_dir}{: path to the FreeSurfer data, this expects a verywise structure.}

\item{hemi}{: (default = "lh") hemisphere.}

\item{measure}{: (default = \code{gsub("vw_", "", all.vars(formula)[1])}) vertex-wise measure
this should be the same as specified in formula}

\item{fwhm}{: (default = 10) full-width half maximum value}

\item{target}{: (default = "fsaverage") template on which to register vertex-wise data.}

\item{n_cores}{: (default = 1) number of cores for parallel processing.}

\item{model}{: (default = \code{"lme4::lmer"}) # "stats::lm"}
}
\value{
A list of file-backed matrices containig pooled coefficients, SEs,
t- and p- values and residuals.
}
\description{
This is is the main function in v0 of verywise runs main analyses.
}
\author{
Serena Defina, 2024.
}
