% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_lmm.R
\name{run_vw_lmm}
\alias{run_vw_lmm}
\title{Run vertex-wise linear mixed model using \code{lme4::lmer()}}
\usage{
run_vw_lmm(
  formula,
  pheno,
  subj_dir,
  outp_dir = NULL,
  hemi = c("both", "lh", "rh"),
  seed = 3108,
  n_cores = 1,
  FS_HOME = Sys.getenv("FREESURFER_HOME"),
  folder_id = "folder_id",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{formula}{: model formula object (this should specify a LME model)}

\item{pheno}{: the phenotype data object (already loaded in the global environment) or
a string containing a file path. Supported file extensions are: rds, csv, txt and sav.}

\item{subj_dir}{: path to the FreeSurfer data, this expects a verywise structure.}

\item{outp_dir}{: output path, where do you want results to be stored. If none is
provided by the user, a "results" sub-directory will be created inside \code{subj_dir}.}

\item{hemi}{: (default = "both") which hemispheres to run.}

\item{seed}{: (default = 3108) random seed.}

\item{n_cores}{: (default = 1) number of cores for parallel processing.}

\item{FS_HOME}{: FreeSurfer directory, i.e. \code{$FREESURFER_HOME}.}

\item{folder_id}{: (default = "folder_id") the name of the column in pheno that
contains the directory names of the input neuroimaging data (e.g. "sub-10_ses-T1").}

\item{verbose}{: (default = TRUE)}

\item{...}{
  Arguments passed on to \code{\link[=hemi_vw_lmm]{hemi_vw_lmm}}
  \describe{
    \item{\code{fwhm}}{: (default = 10) full-width half maximum value}
    \item{\code{target}}{: (default = "fsaverage") template on which to register vertex-wise data.
The following values are accepted:
 * fsaverage (default) = 163842 vertices (highest resolution),
 * fsaverage6 = 40962 vertices,
 * fsaverage5 = 10242 vertices,
 * fsaverage4 = 2562 vertices,
 * fsaverage3 = 642 vertices
Note that, at the moment, these are only used to downsample the brain map, for faster
model tuning. `verywise` expects the input data to be always registered on the "fsaverage"
template and the final analyses should also be run using `target = "fsaverage"`
to avoid (small) imprecisions in vertex registration and smoothing.}
    \item{\code{apply_cortical_mask}}{: (default = TRUE) remove vertices that are not on the cortex.}
    \item{\code{save_ss}}{: (default = TRUE) save the super-subject matrix as an rds file for
quicker processing in the future.}
    \item{\code{model}}{: (default = \code{"lme4::lmer"}) # "stats::lm"}
    \item{\code{use_model_template}}{: pre-compile the model?}
  }}
}
\value{
A list of file-backed matrices containing pooled coefficients, SEs,
t- and p- values and residuals.
}
\description{
This is is the main function in v0 of \code{verywise}.
It checks the user inputs and runs the \code{\link{single_lmm}} function across all vertices.
}
\author{
Serena Defina, 2024.
}
