% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_meta.R
\name{run_vw_meta}
\alias{run_vw_meta}
\title{Vertex-wise Random-Effects Meta-Analysis Across Studies}
\usage{
run_vw_meta(
  term,
  hemi = c("lh", "rh"),
  measure = "area",
  study_names,
  study_weights = NULL,
  res_dirs,
  outp_dir = NULL,
  mtc = "fdr",
  fs_template = "fsaverage",
  seed = 3108,
  n_cores = 1,
  chunk_size = 1000,
  FS_HOME = Sys.getenv("FREESURFER_HOME"),
  mcz_thr = 30,
  cwp_thr = 0.025,
  save_optional_cluster_info = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{term}{Character. Name of the model term to visualize (matches entries in \code{stack_names.txt}).}

\item{hemi}{Character string specifying which hemisphere to analyze.
Options: \code{"lh"} (left hemisphere: default), \code{"rh"} (right hemisphere).}

\item{measure}{Character. Surface measure, e.g. \code{'area'}, \code{'thickness'}, \code{'volume'}. Defaults to \code{'area'}.}

\item{study_names}{Character vector of study names (must match the length of \code{res_dirs}).}

\item{study_weights}{Numeric vector of study weights (e.g. sample size)}

\item{res_dirs}{Character vector. Path to the directories containing vertex-wise result files (\verb{*.mgh}) of each study.}

\item{outp_dir}{Character string specifying the output directory for results.
If \code{NULL} (default), creates a "verywise_results" sub-directory in the
current working directory (not recommended).}

\item{mtc}{Character string: multiple testing correction strategy.
Options: \code{"fdr"} (False Discovery Rate: default), \code{"fs"} (FreeSurfer cluster correction).}

\item{fs_template}{Character string specifying the FreeSurfer template for
vertex registration. Options:
\itemize{
\item \code{"fsaverage"} (default) = 163842 vertices (highest resolution),
\item \code{"fsaverage6"} = 40962 vertices,
\item \code{"fsaverage5"} = 10242 vertices,
\item \code{"fsaverage4"} = 2562 vertices,
\item \code{"fsaverage3"} = 642 vertices
}
Note that lower resolutions should be only used to downsample the brain
map, for faster model tuning. The final analyses should also run using
\code{fs_template = "fsaverage"} to avoid (small) imprecisions in vertex
registration and smoothing.}

\item{seed}{Integer specifying the random seed for reproducibility
Default: 3108.}

\item{n_cores}{Integer specifying the number of CPU cores for parallel
processing.
Default: 1.}

\item{chunk_size}{Integer specifying the number of vertices processed per
chunk in parallel operations. Larger values use more memory but may be
faster.
Default: 1000.}

\item{FS_HOME}{Character string specifying the FreeSurfer home directory.
Defaults to \code{FREESURFER_HOME} environment variable.}

\item{mcz_thr}{Numeric value for the Monte Carlo simulation threshold.
Any of the following are accepted (equivalent values are separated by \code{/}):
\itemize{
\item 13 / 1.3 / 0.05,
\item 20 / 2.0 / 0.01,
\item 23 / 2.3 / 0.005,
\item 30 / 3.0 / 0.001, \* default
\item 33 / 3.3 / 0.0005,
\item 40 / 4.0 / 0.0001.
}}

\item{cwp_thr}{Numeric value for cluster-wise p-value threshold (on top of all
corrections). Set this should be set to \code{0.025} when both hemispheres are analyzed,
and \code{0.05} for single hemisphere analyses.
Default: \code{0.025}.}

\item{save_optional_cluster_info}{Logical indicating whether to save additional
output form \code{mri_surfcluster} call. See \code{\link{compute_clusters}}
for details. Default: \code{FALSE}.}

\item{verbose}{Logical. verbose execution (default: TRUE)}
}
\value{
A named list with three \code{FBM} objects:
\describe{
\item{coef}{Meta-analytic effect size estimates at each vertex.}
\item{se}{Meta-analytic standard errors at each vertex.}
\item{p}{Meta-analytic p-values at each vertex.}
}
Additionally, these are exported as MGH files in the working directory.
}
\description{
This function runs a vertex-wise random-effects meta-analysis across multiple
studies (sites or cohorts). Is expects individual study results as they are
outputted but \code{verywise::run_vw_lmm} or by \code{QDECR}.
}
\details{
For each vertex, the function loads the effect size and standard error from each study, computes the variance,
and runs a random-effects meta-analysis using \code{\link[metafor]{rma}}. Results are saved as
Filebacked Big Matrices (FBM) and as MGH files for downstream neuroimaging analysis.
}
\examples{
\dontshow{if (dir.exists("study1/results")) withAutoprint(\{ # examplesIf}
run_vw_meta(
  term = "age",
  hemi = "lh",
  measure = "area",
  study_names = c("Study1", "Study2"),
  res_dirs = c("study1/results", "study2/results")
)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[metafor]{rma}}, \code{\link[bigstatsr]{FBM}}
}
