% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_meta.R
\name{run_vw_meta}
\alias{run_vw_meta}
\title{Vertex-wise Random-Effects Meta-Analysis Across Studies}
\usage{
run_vw_meta(
  stack,
  hemi,
  result_dirs,
  study_names,
  fs_template = "fsaverage",
  n_cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{stack}{: Character or numeric identifier for the fixed term to meta-analyise.}

\item{hemi}{: Character string, hemisphere to analyze (\code{"lh"} or \code{"rh"}).}

\item{result_dirs}{: Character vector of directories containing the results files of each study.}

\item{study_names}{: Character vector of study names (must match the length of \code{result_dirs}).}

\item{fs_template}{: Character string specifying the FreeSurfer template surface.
The following values are accepted:
\itemize{
\item fsaverage (default) = 163842 vertices (highest resolution),
\item fsaverage6 = 40962 vertices,
\item fsaverage5 = 10242 vertices,
\item fsaverage4 = 2562 vertices,
\item fsaverage3 = 642 vertices
}}

\item{n_cores}{: Integer, number of CPU cores to use for parallel processing (default: 1).}

\item{verbose}{: Logical, verbose execution (default: TRUE)}
}
\value{
A named list with three \code{FBM} objects:
\describe{
\item{coef}{Meta-analytic effect size estimates at each vertex.}
\item{se}{Meta-analytic standard errors at each vertex.}
\item{p}{Meta-analytic p-values at each vertex.}
}
Additionally, these are exported as MGH files in the working directory.
}
\description{
This function runs a vertex-wise random-effects meta-analysis across multiple
studies (sites or cohorts). Is expects individual study results as they are
outputted but \code{verywise::run_vw_lmm} or by \code{QDECR}.
}
\details{
For each vertex, the function loads the effect size and standard error from each study, computes the variance,
and runs a random-effects meta-analysis using \code{\link[metafor]{rma}}. Results are saved as
Filebacked Big Matrices (FBM) and as MGH files for downstream neuroimaging analysis.
}
\examples{
\dontrun{
run_vw_meta(
  stack = 1,
  hemi = "lh",
  result_dirs = c("study1/results", "study2/results"),
  study_names = c("Study1", "Study2"),
  fs_template = "fsaverage",
  n_cores = 4
)
}

}
\seealso{
\code{\link[metafor]{rma}}, \code{\link[bigstatsr]{FBM}}
}
