% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataset.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Simulate longitudinal brain surface dataset with phenotype}
\usage{
simulate_dataset(
  path,
  data_structure = list(cohort1 = list(sessions = c("01", "02", "03"), n_subjects = 10),
    cohort2 = list(sessions = c("01", "02"), n_subjects = 20)),
  simulate_association = NULL,
  overwrite = TRUE,
  seed = 31081996,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{path}{Where should the dataset be created.}

\item{data_structure}{A nested list, with top level determining the cohort/
dataset/site. Each site is itself a list with two items: \code{"sessions"}: a
vector of session names/numbers; and \code{"n_subjects"}: an integer indicating
the number of subjects.}

\item{simulate_association}{(default = NULL) simulate an association in the
format \code{"[beta] * [variable name]"}. For example \code{"0.05 * age"}.
This is by default isolated to three regions:
the superior temporal gyrus, precentral gyrus and middle temporal gyrus.}

\item{overwrite}{(default = TRUE) whether phenotype file should be overwitten.}

\item{seed}{(default = 3108) seed used for randomization.}

\item{verbose}{(deafult = TRUE)}

\item{...}{Other arguments to be passed to \code{\link{simulate_freesurfer_data}}}
}
\value{
NULL. Files are written to `path`.
}
\description{
This function generates synthetic longitudinal datasets for
multiple sites/cohorts (as well as multiple timepoints/sessions per subject).
It generates:
\itemize{
\item the brain surface data in FreeSurfer format (.mgh files) organised in a
 verywise folder structure (see vignettes).
\item a mathing \code{pheno} dataframe with participant sex and age mock-data,
 saved as "phenotype.csv" file in the \code{path} directory.
}
}
\seealso{
\code{\link{simulate_freesurfer_data}}, \code{\link{simulate_long_pheno_data}}
}
\author{
Serena Defina, 2024.
}
