% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataset.R
\name{simulate_freesurfer_data}
\alias{simulate_freesurfer_data}
\title{Simulate FreeSurfer data}
\usage{
simulate_freesurfer_data(
  path,
  vw_resolution = 163842,
  data_structure = list(cohort1 = list(sessions = c("01", "02"), n_subjects = 100),
    cohort2 = list(sessions = c("01", "02"), n_subjects = 150)),
  measure = "thickness",
  hemi = "lh",
  fwhmc = "fwhm10",
  target = "fsaverage",
  vw_mean = 6.5,
  vw_sd = 0.5,
  seed = 31081996
)
}
\arguments{
\item{path}{: Where should the data be created.}

\item{vw_resolution}{: (default = 163842) data dimension (how many vertices)}

\item{data_structure}{: A nested list, with top level determining the cohort/
dataset/site. Each site is itself a list with two items: \code{"sessions"}: a
vector of session names/numbers; and \code{"n_subjects"}: aninteger indicating
the number of subjects.}

\item{measure}{: (default = "thickness"), measure, used in file names.}

\item{hemi}{: (default = "lh") hemisphere, used in file names.}

\item{fwhmc}{: (default = "fwhm10") full-width half maximum value, used in
file names.}

\item{target}{: (default = "fsaverage"), used in file names.}

\item{vw_mean}{: (default = 6.5) mean of the simulated vertex-wise data.}

\item{vw_sd}{: (default = 0.5) standard deviation of the simulated vertex-wise data.}

\item{seed}{: (default = 31081996) seed used for randomization.}
}
\description{
Simulating FreeSurfer data for multiple cohorts/sites and multiple
timepoints/sessions. The funtions emulated the folder structure output obtained
by calling recon_all. This follows the folder names "sub-", participant ID,
"_ses-", session ID.
}
