% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataset.R
\name{simulate_freesurfer_data}
\alias{simulate_freesurfer_data}
\title{Simulate FreeSurfer data}
\usage{
simulate_freesurfer_data(
  path,
  data_structure = list(cohort1 = list(sessions = c("01", "02"), n_subjects = 100),
    cohort2 = list(sessions = c("01", "02"), n_subjects = 150)),
  measure = "thickness",
  hemi = "lh",
  fwhmc = "fwhm10",
  fs_template = "fsaverage",
  vw_mean = 6.5,
  vw_sd = 0.5,
  simulate_association = NULL,
  seed = 3108,
  verbose = TRUE
)
}
\arguments{
\item{path}{Where should the data be created.}

\item{data_structure}{A nested list, with top level determining the cohort/
dataset/site. Each site is itself a list with two items: \code{"sessions"}: a
vector of session names/numbers; and \code{"n_subjects"}: an integer indicating
the number of subjects.}

\item{measure}{(default = "thickness"), measure, used in file names.}

\item{hemi}{(default = "lh") hemisphere, used in file names.}

\item{fwhmc}{(default = "fwhm10") full-width half maximum value, used in
file names.}

\item{fs_template}{Character string specifying the FreeSurfer template for
vertex registration. Options:
\itemize{
\item \code{"fsaverage"} (default) = 163842 vertices (highest resolution),
\item \code{"fsaverage6"} = 40962 vertices,
\item \code{"fsaverage5"} = 10242 vertices,
\item \code{"fsaverage4"} = 2562 vertices,
\item \code{"fsaverage3"} = 642 vertices
}}

\item{vw_mean}{(default = 6.5) mean of the simulated vertex-wise data.}

\item{vw_sd}{(default = 0.5) standard deviation of the simulated vertex-wise data.}

\item{simulate_association}{(default = NULL) simulate an association in the
format \code{"[beta] * [variable name]"}. For example \code{"0.05 * age"}.
This is by default isolated to three regions:
the superior temporal gyrus, precentral gyrus and middle temporal gyrus.}

\item{seed}{(default = 3108) seed used for randomization.}

\item{verbose}{(dafault = TRUE) verbosity.}
}
\description{
Simulating FreeSurfer data for multiple cohorts/sites and multiple
timepoints/sessions. This function emulates the folder structure output obtained
by calling FreeSurfer recon_all command. This follows the folder names "sub-", participant ID,
"_ses-", session ID.
}
\author{
Serena Defina, 2024.
}
