% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataset.R
\name{simulate_long_pheno_data}
\alias{simulate_long_pheno_data}
\title{Simulate phenotype data}
\usage{
simulate_long_pheno_data(
  data_structure = list(cohort1 = list(sessions = c("01", "02"), n_subjects = 100),
    cohort2 = list(sessions = c("01", "02"), n_subjects = 150)),
  seed = 31081996
)
}
\arguments{
\item{data_structure}{A nested list, with top level determining the cohort/
dataset/site. Each site is itself a list with two items: \code{"sessions"}: a
vector of session names/numbers; and \code{"n_subjects"}: an integer indicating
the number of subjects.}

\item{seed}{(default = 31081996) seed used for randomization.}
}
\value{
A dataframe (in long format) with id, sex and age data.
}
\description{
Simulating phenotype data (in the long format) for multiple cohorts/sites and
multiple timepoints/sessions.
}
\author{
Serena Defina, 2024.
}
