% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_lmm.R
\name{single_lmm}
\alias{single_lmm}
\title{Run a single \code{lme4::lmer} model and extract stats}
\usage{
single_lmm(imp, y, formula)
}
\arguments{
\item{imp}{: The phenotype dataset (in verywise format).}

\item{y}{: A vector of outcome values (i.e. a single vertex from the
supersubject matrix).}

\item{formula}{: R formula describing the linear mixed model (using \code{lme4} notation)}
}
\value{
A list with two elements:
\enumerate{
\item \code{"stats"}: a dataframe with estimates, SEs and p-values for each
fixed effect term)
\item \code{"resid"}: a vector of residuals for the given model.
\item \code{"df"}: residual degrees of freedom for the give model.
}
}
\description{
Run a single \code{lme4::lmer} model and extract stats
}
\details{
No additional parameters are currently passed to the \code{lme4::lmer} call
P-values are estimated using \code{car::Anova(., type = 3)} at the moment
}
