% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vw_lmm.R
\name{single_lmm}
\alias{single_lmm}
\title{Run a single \code{lme4::lmer} model and extract stats}
\usage{
single_lmm(imp, y, formula, model_template = NULL, pvalues = TRUE)
}
\arguments{
\item{imp}{: The phenotype dataset (in verywise format).}

\item{y}{: A vector of outcome values (i.e. a single vertex from the
supersubject matrix).}

\item{formula}{: R formula describing the linear mixed model (using \code{lme4} notation)}

\item{model_template}{: (default = NULL) `single_lmm` can leverage an "update"-based
workflow to minimize repeated parsing and model construction overhead.}

\item{pvalues}{: (default : TRUE) whether to include p-values computed using the
*t-as-z* approach (see Details).}
}
\value{
A list with two elements:
\enumerate{
\item \code{"stats"}: a dataframe with estimates, SEs and p-values for each
fixed effect term)
\item \code{"resid"}: a vector of residuals for the given model.
}
}
\description{
Run a single \code{lme4::lmer} model and extract stats
}
\details{
No additional parameters are currently passed to the \code{lme4::lmer} call
P-values are estimated using the *t-as-z* approach at the moment. This is known
to the anti-conservative for small sample sizes. However we preferred a
relatively lenient (and computationally inexpensive) solution at this stage.
We will be addressing Type I error mores strictly at the cluster forming stage.
}
