% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vw_pool}
\alias{vw_pool}
\title{Pool model output across imputed datasets}
\usage{
vw_pool(qhat, se, m, fe_n)
}
\arguments{
\item{qhat}{: a matrix with estimates for the fixed effects}

\item{se}{: a matrix with standard errors of the fixed effects}

\item{m}{: number of imputed datasets}

\item{fe_n}{: number of fixed effect terms}
}
\value{
A list containing the pooled coefficients, SEs, t- and p- values.
}
\description{
This is a rewrite of mice::pool wrote by Sander to allow for multiple outcomes.
This was based on old distributions of mice and may be soon rewritten to take
broom and tidy models formats (like mice > 3 does). This is currently used inside
\code{\link{run_vw_lmm}}
}
\author{
Sander Lamballais, 2018.
}
