% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vw_pool.R
\name{vw_pool}
\alias{vw_pool}
\title{Pool \code{lme4::lmer} model output across imputed datasets}
\usage{
vw_pool(out_stats, m)
}
\arguments{
\item{out_stats}{: Output of \code{\link{single_lmm}}, i.e.: a list with two
elements:
\enumerate{
\item \code{"stats"}: a dataframe with estimates, SEs and p-values for each
fixed effect term)
\item \code{"resid"}: a vector of residuals for the given model.
}}

\item{m}{: Number of imputed dataset (to avoid recomputing it)}
}
\value{
A list containing the pooled coefficients, SEs, t- and p- values.
}
\description{
This function combines estimates, standard errors and p-values across imputed
datasets, for a single model (i.g., one vertex).
The function was largely taken from \code{mice::pool()} and
\code{mice::summary.mipo()} code. It averages the estimates of the
complete data model, and computes relevant statistics, following Rubin's
rules (Rubin, 1987, p. 76).

P-values are estimated using the *t-as-z* approach at the moment. This is known
to the anti-conservative for small sample sizes. However we preferred a
relatively lenient (and computationally inexpensive) solution at this stage.
We will be addressing Type I error mores strictly at the cluster forming stage.

The residuals of the model are currently simply averaged across imputed
datasets, for lack of a better idea of how to combine them.
}
\note{
Used inside \code{\link{run_vw_lmm}}.
}
\references{
Rubin, D.B. (1987). \emph{Multiple Imputation for Nonresponse in Surveys}.
New York: John Wiley and Sons.
}
\author{
Serena Defina, 2024.
}
